+++ src/Controller/CategoryController.php
@@ -41,16 +41,7 @@ class CategoryController extends AbstractController
         $this->entityManager->persist($category);
         $this->entityManager->flush();
 
-        return $this->json([
-            'status' => true,
-            'message' => [
-                'class' => 'CategoryController',
-                'method' => 'createCategory',
-                'content' => [
-                    'Create_Row' => $category->getId()
-                ]
-            ]
-        ]);
+        return new JsonResponse(['status' => true, 'message' => ['Create_Row' => $category->getId()]]);
     }
 
     #[Route('/api/category/{id}', name: 'update_category', methods: ['PATCH'], requirements: ['id' => '\d+'])]
@@ -58,73 +49,49 @@ class CategoryController extends AbstractController
     public function updateCategory(CategoryUpdateIn $categoryUpdateIn, $id): JsonResponse
     {
         $localCategory = $this->categoryRepository->find($id);
+        if(!$localCategory) {
+            return $this->json(['status' => false, 'message' => ['Update_Row' => "Row_" . $id . "_not_found"]]);
+        }
         $category = $this->autoMapper->mapToObject($categoryUpdateIn, $localCategory);
         $this->entityManager->flush();
 
-        return $this->json([
-            'status' => true,
-            'message' => [
-                'class' => 'CategoryController', 
-                'method' => 'updateCategory', 
-                'content' => [
-                    'Update_Row' => $category->getId()
-                ]
-            ]
-        ]);
+        return new JsonResponse(['status' => true, 'message' => ['Update_Row' => $category->getId()]]);
     }
 
     #[Route('/api/category/{id}', name: 'get_category_by_id', methods: ['GET'], requirements: ['id' => '\d+'])]
     public function getCategoryById($id): JsonResponse
     {
         $localCategory = $this->categoryRepository->find($id);
+        if(!$localCategory) {
+            return $this->json(['status' => false, 'message' => ['Category' => "Row_" . $id . "_not_found"]]);
+        }
         $category = $this->autoMapper->map($localCategory, CategoryOut::class);
 
-        return $this->json([
-            'status' => true,
-            'message' => [
-                'class' => 'CategoryController', 
-                'method' => 'getCategoryById',
-                'content' => [
-                    'category' => $category
-                ]
-            ]
-        ]);
+        return new JsonResponse(['status' => true, 'message' => ['Category' => $category]]);
     }
 
     #[Route('/api/category/list', name: 'get_category_list', methods: ['GET'])]
     public function getCategoryList(): JsonResponse
     {
-        $localCategory = $this->categoryRepository->findAll();
-        $category = $this->autoMapper->mapMultiple($localCategory, CategoryListOut::class);
+        $localCategories = $this->categoryRepository->findAll();
+        if(!$localCategories) {
+            return $this->json(['status' => false, 'message' => ['Categories' => "Rows_not_found"]]);
+        }
+        $categories = $this->autoMapper->mapMultiple($localCategories, CategoryListOut::class);
         
-        return $this->json([
-            'status' => true,
-            'message' => [
-                'class' => 'CategoryController', 
-                'method' => 'getCategoryList',
-                'content' => [
-                    'category' => $category
-                ]
-            ]
-        ]);
+        return new JsonResponse(['status' => true, 'message' => ['Categories' => $categories]]);
     }
 
     #[Route('/api/category/{id}', name: 'delete_category', methods: ['DELETE'], requirements: ['id' => '\d+'])]
     public function deleteCategory($id): JsonResponse
     {
         $localCategory = $this->categoryRepository->find($id);
+        if(!$localCategory) {
+            return $this->json(['status' => false, 'message' => ['Delete_Row' => "Row_" . $id . "_not_found"]]);
+        }
         $this->entityManager->remove($localCategory);
         $this->entityManager->flush();
 
-        return $this->json([
-            'status' => true,
-            'message' => [
-                'class' => 'CategoryController',
-                'method' => 'deleteCategory',
-                'content' => [
-                    'Delete_Row' => $id
-                ]
-            ]
-        ]);
+        return new JsonResponse(['status' => true, 'message' => ['Delete_Row' => $id]]);
     }
 }
